own envConfig per each entrypoint; added schema UTs; changed payment amount type to int64 to provide simplicity during convertions; data moved to constants; refactored workjer entrypoint to have a separate worker entity with own deps; moved repo interface to ports; events as part of domain; updateAt field for payments DB entity;

+++ internal/adapters/3rdparty/dto.go
@@ -9,11 +9,11 @@ type AddPaymentMethodResponse struct {
 }
 
 type CreatePaymentWithTokenRequest struct {
-	OrderID       string  `json:"order_id"`
-	Amount        float64 `json:"amount"`
-	Currency      string  `json:"currency"`
-	CustomerID    string  `json:"customer_id"`
-	CustomerToken string  `json:"customer_token"`
+	OrderID       string `json:"order_id"`
+	Amount        int64  `json:"amount"`
+	Currency      string `json:"currency"`
+	CustomerID    string `json:"customer_id"`
+	CustomerToken string `json:"customer_token"`
 }
 
 type CreatePaymentWithTokenResponse struct {
